moving trace to client-go
