+++ hack/import-restrictions.yaml
@@ -60,7 +60,6 @@
   allowedImports:
   - k8s.io/api
   - k8s.io/apimachinery
-  - k8s.io/apiserver/pkg/util/trace
   - k8s.io/client-go
   - k8s.io/klog
 
@@ -83,7 +82,6 @@
   - "./vendor/k8s.io/client-go/tools/remotecommand"
   allowedImports:
   - k8s.io/apimachinery
-  - k8s.io/apiserver/pkg/util/trace
   - k8s.io/client-go
   - k8s.io/klog
