+++ pkg/util/iptables/BUILD
@@ -20,7 +20,7 @@ go_library(
         "//pkg/util/dbus:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/sets:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/version:go_default_library",
-        "//staging/src/k8s.io/apiserver/pkg/util/trace:go_default_library",
+        "//staging/src/k8s.io/client-go/util/trace:go_default_library",
         "//vendor/github.com/godbus/dbus:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
         "//vendor/k8s.io/utils/exec:go_default_library",
