+++ staging/src/BUILD
@@ -104,7 +104,6 @@ filegroup(
         "//staging/src/k8s.io/apiserver/pkg/util/globalflag:all-srcs",
         "//staging/src/k8s.io/apiserver/pkg/util/openapi:all-srcs",
         "//staging/src/k8s.io/apiserver/pkg/util/proxy:all-srcs",
-        "//staging/src/k8s.io/apiserver/pkg/util/trace:all-srcs",
         "//staging/src/k8s.io/apiserver/pkg/util/webhook:all-srcs",
         "//staging/src/k8s.io/apiserver/pkg/util/wsstream:all-srcs",
         "//staging/src/k8s.io/apiserver/plugin/pkg/audit:all-srcs",
@@ -188,6 +187,7 @@ filegroup(
         "//staging/src/k8s.io/client-go/util/jsonpath:all-srcs",
         "//staging/src/k8s.io/client-go/util/retry:all-srcs",
         "//staging/src/k8s.io/client-go/util/testing:all-srcs",
+        "//staging/src/k8s.io/client-go/util/trace:all-srcs",
         "//staging/src/k8s.io/client-go/util/workqueue:all-srcs",
         "//staging/src/k8s.io/cloud-provider:all-srcs",
         "//staging/src/k8s.io/cluster-bootstrap/token/api:all-srcs",
