+++ staging/src/k8s.io/apiserver/pkg/endpoints/handlers/delete.go
@@ -35,7 +35,7 @@ import (
 	"k8s.io/apiserver/pkg/registry/rest"
 	"k8s.io/apiserver/pkg/util/dryrun"
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
-	utiltrace "k8s.io/apiserver/pkg/util/trace"
+	utiltrace "k8s.io/client-go/util/trace"
 )
 
 // DeleteResource returns a function that will handle a resource deletion
