+++ staging/src/k8s.io/apiserver/pkg/endpoints/handlers/get.go
@@ -37,7 +37,7 @@ import (
 	"k8s.io/apiserver/pkg/endpoints/metrics"
 	"k8s.io/apiserver/pkg/endpoints/request"
 	"k8s.io/apiserver/pkg/registry/rest"
-	utiltrace "k8s.io/apiserver/pkg/util/trace"
+	utiltrace "k8s.io/client-go/util/trace"
 )
 
 // getterFunc performs a get request with the given context and object name. The request
