+++ staging/src/k8s.io/apiserver/pkg/endpoints/handlers/patch.go
@@ -44,7 +44,7 @@ import (
 	"k8s.io/apiserver/pkg/registry/rest"
 	"k8s.io/apiserver/pkg/util/dryrun"
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
-	utiltrace "k8s.io/apiserver/pkg/util/trace"
+	utiltrace "k8s.io/client-go/util/trace"
 )
 
 // PatchResource returns a function that will handle a resource patch.
