+++ staging/src/k8s.io/apiserver/pkg/endpoints/handlers/rest.go
@@ -41,7 +41,7 @@ import (
 	"k8s.io/apiserver/pkg/endpoints/metrics"
 	"k8s.io/apiserver/pkg/endpoints/request"
 	"k8s.io/apiserver/pkg/registry/rest"
-	utiltrace "k8s.io/apiserver/pkg/util/trace"
+	utiltrace "k8s.io/client-go/util/trace"
 	openapiproto "k8s.io/kube-openapi/pkg/util/proto"
 )
