+++ staging/src/k8s.io/apiserver/pkg/endpoints/handlers/rest_test.go
@@ -43,7 +43,7 @@ import (
 	examplev1 "k8s.io/apiserver/pkg/apis/example/v1"
 	"k8s.io/apiserver/pkg/endpoints/request"
 	"k8s.io/apiserver/pkg/registry/rest"
-	utiltrace "k8s.io/apiserver/pkg/util/trace"
+	utiltrace "k8s.io/client-go/util/trace"
 )
 
 var (
