+++ staging/src/k8s.io/apiserver/pkg/endpoints/handlers/update.go
@@ -38,7 +38,7 @@ import (
 	"k8s.io/apiserver/pkg/registry/rest"
 	"k8s.io/apiserver/pkg/util/dryrun"
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
-	utiltrace "k8s.io/apiserver/pkg/util/trace"
+	utiltrace "k8s.io/client-go/util/trace"
 )
 
 // UpdateResource returns a function that will handle a resource update
