+++ staging/src/k8s.io/apiserver/pkg/storage/cacher/BUILD
@@ -26,8 +26,8 @@ go_library(
         "//staging/src/k8s.io/apiserver/pkg/features:go_default_library",
         "//staging/src/k8s.io/apiserver/pkg/storage:go_default_library",
         "//staging/src/k8s.io/apiserver/pkg/util/feature:go_default_library",
-        "//staging/src/k8s.io/apiserver/pkg/util/trace:go_default_library",
         "//staging/src/k8s.io/client-go/tools/cache:go_default_library",
+        "//staging/src/k8s.io/client-go/util/trace:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
     ],
 )
