+++ staging/src/k8s.io/apiserver/pkg/storage/cacher/cacher.go
@@ -39,8 +39,8 @@ import (
 	"k8s.io/apiserver/pkg/features"
 	"k8s.io/apiserver/pkg/storage"
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
-	utiltrace "k8s.io/apiserver/pkg/util/trace"
 	"k8s.io/client-go/tools/cache"
+	utiltrace "k8s.io/client-go/util/trace"
 )
 
 // Config contains the configuration for a given Cache.
