+++ staging/src/k8s.io/apiserver/pkg/storage/cacher/watch_cache.go
@@ -29,8 +29,8 @@ import (
 	"k8s.io/apimachinery/pkg/util/clock"
 	"k8s.io/apimachinery/pkg/watch"
 	"k8s.io/apiserver/pkg/storage"
-	utiltrace "k8s.io/apiserver/pkg/util/trace"
 	"k8s.io/client-go/tools/cache"
+	utiltrace "k8s.io/client-go/util/trace"
 )
 
 const (
