+++ staging/src/k8s.io/apiserver/pkg/storage/etcd3/BUILD
@@ -63,7 +63,7 @@ go_library(
         "//staging/src/k8s.io/apiserver/pkg/storage:go_default_library",
         "//staging/src/k8s.io/apiserver/pkg/storage/etcd:go_default_library",
         "//staging/src/k8s.io/apiserver/pkg/storage/value:go_default_library",
-        "//staging/src/k8s.io/apiserver/pkg/util/trace:go_default_library",
+        "//staging/src/k8s.io/client-go/util/trace:go_default_library",
         "//vendor/github.com/coreos/etcd/clientv3:go_default_library",
         "//vendor/github.com/coreos/etcd/etcdserver/api/v3rpc/rpctypes:go_default_library",
         "//vendor/github.com/coreos/etcd/mvcc/mvccpb:go_default_library",
