+++ staging/src/k8s.io/apiserver/pkg/storage/etcd3/store.go
@@ -39,7 +39,7 @@ import (
 	"k8s.io/apiserver/pkg/storage"
 	"k8s.io/apiserver/pkg/storage/etcd"
 	"k8s.io/apiserver/pkg/storage/value"
-	utiltrace "k8s.io/apiserver/pkg/util/trace"
+	utiltrace "k8s.io/client-go/util/trace"
 )
 
 // authenticatedDataString satisfies the value.Context interface. It uses the key to
